requirements.txt
llama-index-core==0.11.23
llama-parse==0.5.14
llama-index-readers-file==0.3.0
python-dotenv==1.0.1
pandas==2.2.3
langchain==0.3.7
langsmith==0.1.143
langgraph==0.2.50
langchainhub==0.1.21
langchain_community==0.3.7
langchain-anthropic==0.3.0
langchain-openai==0.2.9
streamlit==1.40.1
neo4j==5.26.0
langchain-experimental==0.3.3
plotly==5.24.1
torch==2.5.1
transformers==4.46.3
chromadb==0.5.20